site01/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
gunicorn==21.2.0
Pillow==10.3.0
